"Working prototype" commit.
This commit basically forks off of the apriltags-cpp library because I
couldn't figure out how to get cmake to work properly. Once I have that
resolved I'll remove the src of apriltags-cpp and have that be built and
linked separately. I did have to change the source to get it to compile
though.